+++ lib/baby_squeel/association.rb
@@ -5,17 +5,46 @@ module BabySqueel
     # An Active Record association reflection
     attr_reader :_reflection
 
+    # Specifies the model that the polymorphic
+    # association should join with
+    attr_accessor :_polymorphic_klass
+
     def initialize(parent, reflection)
       @parent = parent
       @_reflection = reflection
-      super(@_reflection.klass)
+
+      # In the case of a polymorphic reflection these
+      # attributes will be set after calling #of
+      unless @_reflection.polymorphic?
+        super @_reflection.klass
+      end
+    end
+
+    def of(klass)
+      unless _reflection.polymorphic?
+        raise PolymorphicSpecificationError.new(_reflection.name, klass)
+      end
+
+      clone.of! klass
+    end
+
+    def of!(klass)
+      self._scope = klass
+      self._table = klass.arel_table
+      self._polymorphic_klass = klass
+      self
+    end
+
+    def needs_polyamorous?
+      _join == Arel::Nodes::OuterJoin || _reflection.polymorphic?
     end
 
     # See JoinExpression#add_to_tree.
     def add_to_tree(hash)
       polyamorous = Polyamorous::Join.new(
         _reflection.name,
-        _join
+        _join,
+        _polymorphic_klass
       )
 
       hash[polyamorous] ||= {}
@@ -50,6 +79,8 @@ module BabySqueel
         super
       elsif _table.is_a? Arel::Nodes::TableAlias
         raise AssociationAliasingError.new(_reflection.name, _table.right)
+      elsif _reflection.polymorphic? && _polymorphic_klass.nil?
+        raise PolymorphicNotSpecifiedError.new(_reflection.name)
       else
         @parent._arel([self, *associations])
       end