+++ lib/baby_squeel/table.rb
@@ -2,16 +2,20 @@ require 'baby_squeel/join_expression'
 
 module BabySqueel
   class Table
-    attr_accessor :_on, :_join, :_table
+    attr_accessor :_on, :_table
+    attr_writer :_join
 
     def initialize(arel_table)
       @_table = arel_table
-      @_join = Arel::Nodes::InnerJoin
     end
 
     # See Arel::Table#[]
     def [](key)
       Nodes::Attribute.new(self, key)
+    end
+
+    def _join
+      @_join ||= Arel::Nodes::InnerJoin
     end
 
     # Alias a table. This is only possible when joining
@@ -79,7 +83,7 @@ module BabySqueel
     def _arel(associations = [])
       if _on
         _join.new(_table, Arel::Nodes::On.new(_on))
-      elsif any_outer_joins?(associations)
+      elsif associations.any?(&:needs_polyamorous?)
         JoinExpression.new(associations)
       elsif associations.any?
         associations.reverse.inject({}) do |names, assoc|
@@ -89,12 +93,6 @@ module BabySqueel
     end
 
     private
-
-    def any_outer_joins?(associations)
-      associations.any? do |assoc|
-        assoc._join == Arel::Nodes::OuterJoin
-      end
-    end
 
     def not_found_error!
       raise NotImplementedError, 'BabySqueel::Table will never raise a NotFoundError'